+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # JEC Exchange Project: Update Release Notes
 
+<a name="jec-exchange-1.1.1"></a>
+## **1.1.1** (2017-07-10)
+
+### Bug Fixes
+
+### Features
+
+- **Gruntfile**: adding support for types generation
+- **security/login:** contains the API for adding login security utilities to a JEC container
+- **security/realms:** contains the API for adding realms to a JEC container
+- **security/crypto:** adds cryptography capabilities to a JEC container
+
+<a name="jec-exchange-1.1.0"></a>
+## **1.1.0** (2017-06-17)
+
+### Bug Fixes
+
+### Features
+
+- This release removes all references to the Node.js API
+
 <a name="jec-exchange-1.0.1"></a>
 ## **1.0.1** (2017-06-05)
 
@@ -7,9 +28,9 @@
 
 ### Features
 
-- This release provides many tests improvements.
+- This release provides many tests improvements
 - Dependencies upgrade
-- **stratup:** contains the API for adding startup scripts to a JEC container
+- **startup:** contains the API for adding startup scripts to a JEC container
 
 <a name="jec-exchange-1.0.0"></a>
 ## **1.0.0** (2017-05-20)